requirements.txt
python-telegram-bot
python-dotenv
pypdf
img2pdf
